Fix up Frontend after Merge
- The frontend is functional again after these hotfixes. All that's left is to do a lot of unit test correcting
