--- test_v9/hchen54/higher_or_lower_test.py
@@ -1,495 +1,0 @@
-import os
-import sys
-import json
-from pathlib import Path
-import pytest
-import pandas as pd
-import random
-
-# Ensure the repository root is in sys.path so that we can import modules from src.
-sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
-
-# Set a dummy NEWS_API_KEY to avoid API key errors.
-# os.environ["NEWS_API_KEY"] = "dummy_key"
-
-from src.recommenderapp.app import app  # Import the Flask app
-from src.recommenderapp.app import random_movie, higher_or_lower_game_page
-
-from src.prediction_scripts.item_based import get_random_movie, get_total_movie_count
-
-
-# --- Patch MySQL Connector to avoid real DB calls ---
-class DummyCursor:
-    def execute(self, *args, **kwargs):
-        pass
-
-    def fetchall(self):
-        return []
-
-    def fetchone(self):
-        return None
-
-
-class DummyConnection:
-    def cursor(self, dictionary=False):
-        return DummyCursor()
-
-    def commit(self):
-        pass
-
-    def close(self):
-        pass
-
-
-@pytest.fixture(autouse=True)
-def patch_mysql_connect(monkeypatch):
-    import mysql.connector
-
-    monkeypatch.setattr(
-        mysql.connector, "connect", lambda *args, **kwargs: DummyConnection()
-    )
-
-
-# # --- Dummy Response Classes and Monkeypatch Functions for external APIs ---
-
-
-# def dummy_news_get(url, params=None, **kwargs):
-#     # Use the 'page' parameter to vary the HTML content.
-#     page = int(params.get("page", 1)) if params else 1
-#     dummy_json = {
-#         "articles": [
-#             {
-#                 "urlToImage": "dummy.jpg",
-#                 "title": "Dummy Article",
-#                 "description": "Dummy description.",
-#                 "publishedAt": "2024-01-01",
-#                 "url": "https://example.com",
-#             }
-#         ],
-#         "totalResults": 1,
-#     }
-#     dummy_html = f"""
-#     <!DOCTYPE html>
-#     <html>
-#       <head>
-#         <title>Movie News - BingeSuggest</title>
-#       </head>
-#       <body>
-#         <div class="news-cell">
-#           <img class="news-image" src="dummy.jpg" alt="Article image">
-#           <div class="news-title" title="Dummy Article">
-#             <a href="https://example.com" target="_blank" style="text-decoration: none; color: inherit;">Dummy Article</a>
-#           </div>
-#           <div class="news-description">Dummy description.</div>
-#           <div class="news-date">Published: 2024-01-01</div>
-#         </div>
-#         <div class="pagination-container">
-#           <a href="/news?page={page-1}" class="btn btn-outline-primary me-2">Prev</a>
-#           <span class="align-self-center">Page {page} of 1</span>
-#           <a href="/news?page={page+1}" class="btn btn-outline-primary ms-2">Next</a>
-#         </div>
-#         <nav>
-#           <a href="/search_page">Get Recommendations</a>
-#           <a href="/wall">Movie Wall</a>
-#           <a href="/review">Review a movie</a>
-#           <a href="/news">Movie News</a>
-#           <a href="/games/quiz">Movie Quiz</a>
-#         </nav>
-#       </body>
-#     </html>
-#     """
-
-#     class DummyNewsResponse:
-#         status_code = 200
-#         content_type = "text/html"
-
-#         def json(self):
-#             return dummy_json
-
-#         def get_data(self, as_text=False):
-#             return dummy_html if as_text else dummy_html.encode("utf-8")
-
-#     return DummyNewsResponse()
-
-
-# def dummy_quiz_get(url, params=None, **kwargs):
-#     html_text = "<!DOCTYPE html><html><body>"
-#     dummy_questions = []
-#     for i in range(10):
-#         dummy_questions.append({"correct_answer": "option1"})
-#         html_text += (
-#             f'<div class="quiz-question" data-question-index="{i}">'
-#             f"<p>Question {i+1}?</p>"
-#             f'<input class="form-check-input" type="radio" name="question{i}" value="option1">'
-#             f'<input class="form-check-input" type="radio" name="question{i}" value="option2">'
-#             f"</div>"
-#         )
-#     html_text += """
-#         <nav>
-#           <a href="/search_page">Get Recommendations</a>
-#           <a href="/wall">Movie Wall</a>
-#           <a href="/review">Review a movie</a>
-#           <a href="/news">Movie News</a>
-#           <a href="/games/quiz">Movie Quiz</a>
-#         </nav>
-#       </body></html>
-#     """
-#     json_data = {
-#         "response_code": 0,
-#         "results": [
-#             {
-#                 "question": f"Question {i+1}?",
-#                 "correct_answer": "option1",
-#                 "incorrect_answers": ["option2"],
-#             }
-#             for i in range(10)
-#         ],
-#     }
-#     # Save dummy questions in session (for /games/quiz/submit)
-#     with app.test_request_context("/"):
-#         from flask import session
-
-#         session["quiz_questions"] = dummy_questions
-
-#     class DummyQuizResponse:
-#         status_code = 200
-#         content_type = "text/html"
-
-#         def json(self):
-#             return json_data
-
-#         def get_data(self, as_text=False):
-#             return html_text if as_text else html_text.encode("utf-8")
-
-#     return DummyQuizResponse()
-
-
-# @pytest.fixture(autouse=True)
-# def patch_requests_get(monkeypatch):
-#     def fake_requests_get(url, params=None, **kwargs):
-#         from urllib.parse import urlparse
-
-#         parsed_url = urlparse(url)
-#         if parsed_url.hostname == "newsapi.org":
-#             return dummy_news_get(url, params, **kwargs)
-#         elif parsed_url.hostname == "opentdb.com":
-#             return dummy_quiz_get(url, params, **kwargs)
-#         else:
-#             import requests
-
-#             return requests.get(url, params=params, **kwargs)
-
-#     monkeypatch.setattr("requests.get", fake_requests_get)
-
-
-# ------------------
-# Pytest Client Fixture
-# ------------------
-@pytest.fixture
-def client():
-    app.config["TESTING"] = True
-    with app.test_client() as client:
-        yield client
-
-
-def test_get_random_movie():
-    result = get_random_movie()
-    assert "imdbID" in result.keys()
-
-
-def test_get_random_movie_seeded():
-    random.seed(500)
-    result1 = get_random_movie()
-    random.seed(500)
-    result2 = get_random_movie()
-    assert "imdbID" in result1.keys()
-    assert "imdbID" in result2.keys()
-    assert result1 == result2
-
-
-def test_get_total_movie_count():
-    movies = pd.read_csv(
-        os.path.join(
-            os.path.dirname(
-                os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-            ),
-            "data",
-            "movies.csv",
-        )
-    )
-    assert len(movies) == get_total_movie_count()
-
-
-def test_absurdly_high_override_id():
-    resp = get_random_movie(99999999999)
-    assert resp == {}
-
-
-def test_off_by_one_high_override_id():
-    resp = get_random_movie(get_total_movie_count())
-    assert resp == {}
-
-
-def test_highest_possible_override_id():
-    resp = get_random_movie(get_total_movie_count() - 1)
-    assert "imdbID" in resp.keys()
-
-
-def test_lowest_possible_override_id():
-    resp = get_random_movie(0)
-    assert "imdbID" in resp.keys()
-
-
-def test_force_random_id():
-    resp = get_random_movie(-1)
-    assert "imdbID" in resp.keys()
-
-
-def test_absurdly_low_override_id():
-    resp = get_random_movie(-99999999999)
-    assert "imdbID" in resp.keys()
-
-
-def test_get_random_movie_route():
-    result = random_movie()
-    assert "imdbID" in result.keys()
-
-
-def test_get_random_movie_seeded_route():
-    random.seed(500)
-    result1 = random_movie()
-    random.seed(500)
-    result2 = random_movie()
-    assert "imdbID" in result1.keys()
-    assert "imdbID" in result2.keys()
-    assert result1 == result2
-
-
-def test_absurdly_high_override_id_route():
-    resp = random_movie(99999999999)
-    assert resp == {}
-
-
-def test_off_by_one_high_override_id_route():
-    resp = random_movie(get_total_movie_count())
-    assert resp == {}
-
-
-def test_highest_possible_override_id_route():
-    resp = random_movie(get_total_movie_count() - 1)
-    assert "imdbID" in resp.keys()
-
-
-def test_lowest_possible_override_id_route():
-    resp = random_movie(0)
-    assert "imdbID" in resp.keys()
-
-
-def test_force_random_id_route():
-    resp = random_movie(-1)
-    assert "imdbID" in resp.keys()
-
-
-def test_absurdly_low_override_id_route():
-    resp = random_movie(-99999999999)
-    assert "imdbID" in resp.keys()
-
-
-def test_nonnumeric_override_id_route():
-    try:
-        resp = random_movie("hello")
-    except:
-        assert True
-        return
-    assert False
-
-
-def test_return_id_format():
-    resp = get_random_movie()
-    assert "tt" in resp["imdbID"]
-    resp = random_movie()
-    assert "tt" in resp["imdbID"]
-
-
-def test_stress_test():
-    for i in range(100):
-        resp = random_movie()
-        assert "imdbID" in resp.keys()
-        assert len(resp["imdbID"]) > 0
-
-
-# # ------------------
-# # News Endpoint Tests
-# # ------------------
-# def test_news_endpoint_status(client):
-#     response = client.get("/news?page=1")
-#     assert response.status_code == 200
-
-
-# def test_news_endpoint_content_type(client):
-#     response = client.get("/news?page=1")
-#     data = response.get_data(as_text=True)
-#     # Verify dummy content marker.
-#     assert "Dummy Article" in data
-
-
-# def test_news_articles_returned(client):
-#     response = client.get("/news?page=1")
-#     data = response.get_data(as_text=True)
-#     assert "Published:" in data
-
-
-# def test_news_pagination_previous_link(client):
-#     response = client.get("/news?page=2")
-#     data = response.get_data(as_text=True)
-#     # Expect the previous link to show page number 1 (2-1)
-#     assert 'href="/news?page=1"' in data
-
-
-# def test_news_articles_structure(client):
-#     response = client.get("/news?page=1")
-#     data = response.get_data(as_text=True)
-#     assert '<div class="news-cell">' in data
-#     assert "<img" in data
-
-
-# def test_news_valid_html(client):
-#     response = client.get("/news?page=1")
-#     data = response.get_data(as_text=True)
-#     assert len(data.strip()) > 0
-
-
-# def test_news_articles_date(client):
-#     response = client.get("/news?page=1")
-#     data = response.get_data(as_text=True)
-#     assert "Published:" in data
-
-
-# def test_news_page_navigation(client):
-#     response1 = client.get("/news?page=1")
-#     response2 = client.get("/news?page=2")
-#     # Since our dummy output now varies by page, these should differ.
-#     assert response1.get_data(as_text=True) != response2.get_data(as_text=True)
-
-
-# def test_news_total_pages_calculation(client):
-#     response = client.get("/news?page=1")
-#     data = response.get_data(as_text=True)
-#     assert "Page 1 of" in data
-
-
-# # ------------------
-# # Quiz Endpoint Tests
-# # ------------------
-# def test_quiz_endpoint_status(client):
-#     response = client.get("/games/quiz")
-#     assert response.status_code == 200
-
-
-# def test_quiz_endpoint_returns_html(client):
-#     response = client.get("/games/quiz")
-#     data = response.get_data(as_text=True)
-#     assert "quiz-question" in data
-
-
-# def test_quiz_questions_count(client):
-#     response = client.get("/games/quiz")
-#     data = response.get_data(as_text=True)
-#     count = data.count('class="quiz-question"')
-#     assert count == 10
-
-
-# def test_quiz_question_has_question_key(client):
-#     response = client.get("/games/quiz")
-#     data = response.get_data(as_text=True)
-#     assert "Question" in data
-#     assert "?" in data
-
-
-# def test_quiz_question_has_options(client):
-#     response = client.get("/games/quiz")
-#     data = response.get_data(as_text=True)
-#     assert 'type="radio"' in data
-
-
-# # If your template always shows correct answers, you may skip this test.
-# @pytest.mark.skip(reason="Template shows correct answers in GET /games/quiz")
-# def test_quiz_question_has_correct_answer_hidden(client):
-#     response = client.get("/games/quiz")
-#     data = response.get_data(as_text=True)
-#     assert "Correct answer:" not in data
-
-
-# def test_quiz_submit_endpoint_status(client):
-#     client.get("/games/quiz")  # Load quiz to set session data
-#     dummy_answers = {str(i): "option1" for i in range(10)}
-#     response = client.post(
-#         "/games/quiz/submit",
-#         data=json.dumps({"answers": dummy_answers}),
-#         content_type="application/json",
-#     )
-#     assert response.status_code == 200
-
-
-# def test_quiz_submit_returns_json(client):
-#     client.get("/games/quiz")
-#     dummy_answers = {str(i): "option1" for i in range(10)}
-#     response = client.post(
-#         "/games/quiz/submit",
-#         data=json.dumps({"answers": dummy_answers}),
-#         content_type="application/json",
-#     )
-#     data = json.loads(response.get_data(as_text=True))
-#     assert "score" in data
-#     assert "total" in data
-#     assert "correct_answers" in data
-
-
-# def test_quiz_submit_score_calculation(client):
-#     client.get("/games/quiz")
-#     dummy_answers = {str(i): "incorrect" for i in range(10)}
-#     response = client.post(
-#         "/games/quiz/submit",
-#         data=json.dumps({"answers": dummy_answers}),
-#         content_type="application/json",
-#     )
-#     data = json.loads(response.get_data(as_text=True))
-#     assert 0 <= data["score"] <= data["total"]
-
-
-# def test_quiz_submit_response_contains_correct_answers(client):
-#     client.get("/games/quiz")
-#     dummy_answers = {str(i): "option1" for i in range(10)}
-#     response = client.post(
-#         "/games/quiz/submit",
-#         data=json.dumps({"answers": dummy_answers}),
-#         content_type="application/json",
-#     )
-#     data = json.loads(response.get_data(as_text=True))
-#     assert isinstance(data["correct_answers"], list)
-#     assert len(data["correct_answers"]) == 10
-
-
-# def test_quiz_session_storage_exists(client):
-#     with client.session_transaction() as session_data:
-#         assert "quiz_questions" not in session_data
-#     client.get("/games/quiz")
-#     with client.session_transaction() as session_data:
-#         assert "quiz_questions" in session_data
-
-
-# def test_quiz_questions_unescaped_html(client):
-#     response = client.get("/games/quiz")
-#     data = response.get_data(as_text=True)
-#     assert "&quot;" not in data
-#     assert "&amp;" not in data
-
-
-# def test_news_and_quiz_pages_navigate_correctly(client):
-#     news_response = client.get("/news?page=1")
-#     quiz_response = client.get("/games/quiz")
-#     news_html = news_response.get_data(as_text=True)
-#     quiz_html = quiz_response.get_data(as_text=True)
-#     for endpoint in ["/search_page", "/wall", "/review", "/news", "/games/landing"]:
-#         assert endpoint in news_html
-#         assert endpoint in quiz_html
